standard-OCL: remove private dependence
 * disable cl_intel.h downloading
 * separate cl_va_memory.h from cl_memory.h
 * separate cl_intel_context.h from cl_context.h
 * fix some undefined errors
