+++ modules/ocl/cl_utils.h
@@ -26,6 +26,9 @@
 #include "ocl/cl_context.h"
 #include "ocl/cl_memory.h"
 #include "ocl/cl_video_buffer.h"
+#if HAVE_LIBDRM
+#include "drm_bo_buffer.h"
+#endif
 
 #define XCAM_CL_IMAGE_ALIGNMENT_X 4
